+++ src/sentry/templatetags/sentry_api.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import
 
 from django import template
-from django.http import HttpRequest
 
 from sentry.auth.access import from_user, NoAccess
 from sentry.api.serializers.base import serialize as serialize_func
@@ -43,20 +42,3 @@ def serialize_detailed_org(context, obj):
     )
 
     return convert_to_json(context)
-
-
-@register.simple_tag
-def get_user_context(request, escape=False):
-    if isinstance(request, HttpRequest):
-        user = getattr(request, 'user', None)
-        result = {'ip_address': request.META['REMOTE_ADDR']}
-        if user and user.is_authenticated():
-            result.update({
-                'email': user.email,
-                'id': user.id,
-            })
-            if user.name:
-                result['name'] = user.name
-    else:
-        result = {}
-    return convert_to_json(result)
